+++ src/Oro/Bundle/EmailBundle/DependencyInjection/OroEmailExtension.php
@@ -7,11 +7,6 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
-/**
- * This is the class that loads and manages your bundle configuration
- *
- * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
- */
 class OroEmailExtension extends Extension
 {
     /**
@@ -37,7 +32,6 @@ class OroEmailExtension extends Extension
         $loader->load('services_api.yml');
         $loader->load('form.yml');
         $loader->load('mass_action.yml');
-        $loader->load('filters.yml');
         $loader->load('commands.yml');
 
         $container->prependExtensionConfig($this->getAlias(), array_intersect_key($config, array_flip(['settings'])));