+++ src/Oro/Bundle/SearchBundle/OroSearchBundle.php
@@ -2,14 +2,17 @@
 
 namespace Oro\Bundle\SearchBundle;
 
+use Oro\Bundle\FilterBundle\DependencyInjection\Compiler\FilterTypesPass;
 use Oro\Bundle\MessageQueueBundle\DependencyInjection\Compiler\AddTopicMetaPass;
 use Oro\Bundle\SearchBundle\Async\Topics;
-use Oro\Bundle\SearchBundle\DependencyInjection\Compiler\FilterTypesPass;
 use Oro\Bundle\SearchBundle\DependencyInjection\Compiler\ListenerExcludeSearchConnectionPass;
 use Oro\Component\DependencyInjection\ExtendedContainerBuilder;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
+/**
+ * The SearchBundle bundle class.
+ */
 class OroSearchBundle extends Bundle
 {
     /** {@inheritdoc} */
@@ -31,6 +34,9 @@ class OroSearchBundle extends Bundle
             ->add(Topics::INDEX_ENTITY, 'Index single entity by id')
         ;
         $container->addCompilerPass($addTopicPass);
-        $container->addCompilerPass(new FilterTypesPass());
+        $container->addCompilerPass(new FilterTypesPass(
+            'oro_search.extension.search_filter_bag',
+            'oro_search.extension.search_filter.filter'
+        ));
     }
 }